Work on conditionally required analyte properties
